backend/requirements.txt
Flask==2.2.5
functions-framework==3.4.0

openai==0.28

PyPDF2==3.0.1
python-docx==1.1.0

pytesseract==0.3.10
Pillow==10.2.0

firebase-admin==6.5.0
google-cloud-firestore==2.13.0
google-auth==2.29.0  # Needed for Firestore auth
